requirements.txt
asgiref==3.4.1
boto3==1.17.105
botocore==1.20.105
click==8.0.1
cmake==3.20.5
colorama==0.4.4
Django==3.2.5
django-cors-headers
whitenoise==5.2.0s
django-cleanup==5.2.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.4
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==20.1.0
importlib-metadata==4.6.1
jmespath==0.10.0
Markdown==3.3.4
numpy==1.21.0
opencv-python-headless==4.2.0.32
opencv-contrib-python-headless==4.5.2.54
Pillow==8.3.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0

